seg/segment_service/requirements.txt
opencv-python
pyyaml
numpy
supervision
flask
ultralytics
Pillow 
scipy 
pyamg 
matplotlib 
scikit-image
